add separate endpoints and first trying to test
