+++ tests/conftest.py
@@ -1,0 +1,16 @@
+import os
+import pytest
+
+#
+from flask import Flask
+from utils.session import SessionController
+
+
+@pytest.fixture
+def client():
+    yield Flask(__name__)
+
+
+@pytest.fixture
+def session():
+    return SessionController()
